+++ crates/hodu_core/src/be_cpu/storage/ops_matrix.rs
@@ -47,10 +47,10 @@ pub fn call_ops_matmul(
     }
 
     // Extract matrix dimensions
-    let m = lhs_shape.dims()[(lhs_ndim - 2) as usize];
-    let k_lhs = lhs_shape.dims()[(lhs_ndim - 1) as usize];
-    let k_rhs = rhs_shape.dims()[(rhs_ndim - 2) as usize];
-    let n = rhs_shape.dims()[(rhs_ndim - 1) as usize];
+    let m = lhs_shape.dims()[lhs_ndim - 2];
+    let k_lhs = lhs_shape.dims()[lhs_ndim - 1];
+    let k_rhs = rhs_shape.dims()[rhs_ndim - 2];
+    let n = rhs_shape.dims()[rhs_ndim - 1];
 
     // Check that inner dimensions match
     if k_lhs != k_rhs {
@@ -67,17 +67,17 @@ pub fn call_ops_matmul(
     let batch_ndim = lhs_batch_ndim.max(rhs_batch_ndim);
 
     // Broadcast batch dimensions
-    let mut batch_shape = Vec::with_capacity(batch_ndim as usize);
+    let mut batch_shape = Vec::with_capacity(batch_ndim);
     for i in 0..batch_ndim {
         let lhs_idx = (lhs_batch_ndim as i32 - batch_ndim as i32 + i as i32) as usize;
         let rhs_idx = (rhs_batch_ndim as i32 - batch_ndim as i32 + i as i32) as usize;
 
-        let lhs_dim = if lhs_idx < lhs_batch_ndim as usize {
+        let lhs_dim = if lhs_idx < lhs_batch_ndim {
             lhs_shape.dims()[lhs_idx]
         } else {
             1
         };
-        let rhs_dim = if rhs_idx < rhs_batch_ndim as usize {
+        let rhs_dim = if rhs_idx < rhs_batch_ndim {
             rhs_shape.dims()[rhs_idx]
         } else {
             1
@@ -106,48 +106,46 @@ pub fn call_ops_matmul(
     // Build metadata array for CPU kernel
     // Layout: num_els, lhs_ndim, rhs_ndim, batch_ndim, lhs_shape, rhs_shape, batch_shape,
     //         lhs_strides, rhs_strides, lhs_offset, rhs_offset, M, K, N
-    let mut metadata: Vec<usize> = Vec::with_capacity(
-        4 + lhs_ndim as usize + rhs_ndim as usize + batch_ndim as usize + lhs_ndim as usize + rhs_ndim as usize + 5,
-    );
+    let mut metadata: Vec<usize> = Vec::with_capacity(4 + lhs_ndim + rhs_ndim + batch_ndim + lhs_ndim + rhs_ndim + 5);
 
-    metadata.push(num_els as usize);
-    metadata.push(lhs_ndim as usize);
-    metadata.push(rhs_ndim as usize);
-    metadata.push(batch_ndim as usize);
+    metadata.push(num_els);
+    metadata.push(lhs_ndim);
+    metadata.push(rhs_ndim);
+    metadata.push(batch_ndim);
 
     // Add lhs shape
     for &dim in lhs_shape.dims() {
-        metadata.push(dim as usize);
+        metadata.push(dim);
     }
 
     // Add rhs shape
     for &dim in rhs_shape.dims() {
-        metadata.push(dim as usize);
+        metadata.push(dim);
     }
 
     // Add batch shape
     for &dim in &batch_shape {
-        metadata.push(dim as usize);
+        metadata.push(dim);
     }
 
     // Add lhs strides
     for &stride in lhs_layout.strides() {
-        metadata.push(stride as usize);
+        metadata.push(stride);
     }
 
     // Add rhs strides
     for &stride in rhs_layout.strides() {
-        metadata.push(stride as usize);
+        metadata.push(stride);
     }
 
     // Add offsets
-    metadata.push(lhs_layout.offset() as usize);
-    metadata.push(rhs_layout.offset() as usize);
+    metadata.push(lhs_layout.offset());
+    metadata.push(rhs_layout.offset());
 
     // Add matrix dimensions M, K, N
-    metadata.push(m as usize);
-    metadata.push(k_lhs as usize);
-    metadata.push(n as usize);
+    metadata.push(m);
+    metadata.push(k_lhs);
+    metadata.push(n);
 
     // Generate kernel name
     let kernel_name = format!("matmul_{}", lhs_storage.dtype());
@@ -156,7 +154,7 @@ pub fn call_ops_matmul(
 
     // Create output storage
     let dtype = lhs_storage.dtype();
-    let mut output = CpuDevice::allocate(output_shape.size() as usize, dtype)?;
+    let mut output = CpuDevice::allocate(output_shape.size(), dtype)?;
 
     // Get raw pointers and call kernel
     macro_rules! call_kernel {
@@ -286,21 +284,21 @@ pub fn call_ops_dot(
     // Layout: M, K, N, lhs_stride_m, lhs_stride_k, rhs_stride_k, rhs_stride_n, lhs_offset, rhs_offset
     let mut metadata: Vec<usize> = Vec::with_capacity(9);
 
-    metadata.push(m as usize);
-    metadata.push(k_lhs as usize);
-    metadata.push(n as usize);
+    metadata.push(m);
+    metadata.push(k_lhs);
+    metadata.push(n);
 
     // Add strides
     let lhs_strides = lhs_layout.strides();
     let rhs_strides = rhs_layout.strides();
-    metadata.push(lhs_strides[0] as usize); // stride for rows
-    metadata.push(lhs_strides[1] as usize); // stride for cols
-    metadata.push(rhs_strides[0] as usize); // stride for rows
-    metadata.push(rhs_strides[1] as usize); // stride for cols
+    metadata.push(lhs_strides[0]); // stride for rows
+    metadata.push(lhs_strides[1]); // stride for cols
+    metadata.push(rhs_strides[0]); // stride for rows
+    metadata.push(rhs_strides[1]); // stride for cols
 
     // Add offsets
-    metadata.push(lhs_layout.offset() as usize);
-    metadata.push(rhs_layout.offset() as usize);
+    metadata.push(lhs_layout.offset());
+    metadata.push(rhs_layout.offset());
 
     // Generate kernel name
     let kernel_name = format!("dot_{}", lhs_storage.dtype());
@@ -309,7 +307,7 @@ pub fn call_ops_dot(
 
     // Create output storage
     let dtype = lhs_storage.dtype();
-    let mut output = CpuDevice::allocate(output_shape.size() as usize, dtype)?;
+    let mut output = CpuDevice::allocate(output_shape.size(), dtype)?;
 
     // Get raw pointers and call kernel
     macro_rules! call_kernel {
